requirements.txt
pandas~=2.2.3
matplotlib~=3.10.0
tensorflow~=2.16.2
numpy~=1.26.4
pillow~=11.1.0
seaborn~=0.13.2
scikit-learn~=1.6.1
